调整税收浮点数

+++ src/server/game/internal/module.go
@@ -29,7 +29,7 @@ type Module struct {
 func (m *Module) OnInit() {
 	m.Skeleton = skeleton
 
-	packageTax = make(map[uint16]uint8)
+	packageTax = make(map[uint16]float64)
 
 	// 中心服务器
 	c4c = NewClient4Center()
